Consolidate text input and organize alloc/free
