+++ flip_social.c
@@ -90,88 +90,6 @@ void flip_social_app_free(FlipSocialApp *app)
         variable_item_list_free(app->variable_item_list_logged_in_settings_wifi);
     }
 
-    // Free Text Input(s)
-    if (app->text_input_logged_out_wifi_settings_ssid)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedOutWifiSettingsSSIDInput);
-        uart_text_input_free(app->text_input_logged_out_wifi_settings_ssid);
-    }
-    if (app->text_input_logged_out_wifi_settings_password)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedOutWifiSettingsPasswordInput);
-        uart_text_input_free(app->text_input_logged_out_wifi_settings_password);
-    }
-    if (app->text_input_logged_out_login_username)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedOutLoginUsernameInput);
-        uart_text_input_free(app->text_input_logged_out_login_username);
-    }
-    if (app->text_input_logged_out_login_password)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedOutLoginPasswordInput);
-        uart_text_input_free(app->text_input_logged_out_login_password);
-    }
-    if (app->text_input_logged_out_register_username)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedOutRegisterUsernameInput);
-        uart_text_input_free(app->text_input_logged_out_register_username);
-    }
-    if (app->text_input_logged_out_register_password)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedOutRegisterPasswordInput);
-        uart_text_input_free(app->text_input_logged_out_register_password);
-    }
-    if (app->text_input_logged_out_register_password_2)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedOutRegisterPassword2Input);
-        uart_text_input_free(app->text_input_logged_out_register_password_2);
-    }
-    if (app->text_input_logged_in_change_password)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedInChangePasswordInput);
-        uart_text_input_free(app->text_input_logged_in_change_password);
-    }
-    if (app->text_input_logged_in_change_bio)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedInChangeBioInput);
-        uart_text_input_free(app->text_input_logged_in_change_bio);
-    }
-    if (app->text_input_logged_in_compose_pre_save_input)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedInComposeAddPreSaveInput);
-        uart_text_input_free(app->text_input_logged_in_compose_pre_save_input);
-    }
-    if (app->text_input_logged_in_wifi_settings_ssid)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedInWifiSettingsSSIDInput);
-        uart_text_input_free(app->text_input_logged_in_wifi_settings_ssid);
-    }
-    if (app->text_input_logged_in_wifi_settings_password)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedInWifiSettingsPasswordInput);
-        uart_text_input_free(app->text_input_logged_in_wifi_settings_password);
-    }
-    if (app->text_input_logged_in_messages_new_message)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedInMessagesNewMessageInput);
-        uart_text_input_free(app->text_input_logged_in_messages_new_message);
-    }
-    if (app->text_input_logged_in_messages_new_message_user_choices)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedInMessagesNewMessageUserChoicesInput);
-        uart_text_input_free(app->text_input_logged_in_messages_new_message_user_choices);
-    }
-    if (app->text_input_logged_in_explore)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedInExploreInput);
-        uart_text_input_free(app->text_input_logged_in_explore);
-    }
-    if (app->text_input_logged_in_message_users)
-    {
-        view_dispatcher_remove_view(app->view_dispatcher, FlipSocialViewLoggedInMessageUsersInput);
-        uart_text_input_free(app->text_input_logged_in_message_users);
-    }
-
     // Free Widget(s)
     if (app->widget_result)
     {
@@ -267,41 +185,6 @@ void flip_social_app_free(FlipSocialApp *app)
     // Free the app structure
     if (app_instance)
         free(app_instance);
-}
-
-void auth_headers_alloc(void)
-{
-    if (!app_instance)
-    {
-        snprintf(auth_headers, sizeof(auth_headers), "{\"Content-Type\":\"application/json\"}");
-        return;
-    }
-
-    if (app_instance->login_username_logged_out && app_instance->login_password_logged_out && strlen(app_instance->login_username_logged_out) > 0 && strlen(app_instance->login_password_logged_out) > 0)
-    {
-        snprintf(auth_headers, sizeof(auth_headers), "{\"Content-Type\":\"application/json\",\"username\":\"%s\",\"password\":\"%s\"}", app_instance->login_username_logged_out, app_instance->login_password_logged_out);
-    }
-    else if (app_instance->login_username_logged_in && app_instance->change_password_logged_in && strlen(app_instance->login_username_logged_in) > 0 && strlen(app_instance->change_password_logged_in) > 0)
-    {
-        snprintf(auth_headers, sizeof(auth_headers), "{\"Content-Type\":\"application/json\",\"username\":\"%s\",\"password\":\"%s\"}", app_instance->login_username_logged_in, app_instance->change_password_logged_in);
-    }
-    else
-    {
-        snprintf(auth_headers, sizeof(auth_headers), "{\"Content-Type\":\"application/json\"}");
-    }
-}
-
-FlipSocialFeedMini *flip_feed_info_alloc(void)
-{
-    FlipSocialFeedMini *feed_info = (FlipSocialFeedMini *)malloc(sizeof(FlipSocialFeedMini));
-    if (!feed_info)
-    {
-        FURI_LOG_E(TAG, "Failed to allocate memory for feed_info");
-        return NULL;
-    }
-    feed_info->count = 0;
-    feed_info->index = 0;
-    return feed_info;
 }
 
 void flip_feed_info_free(void)
